Implement per-frame autoframeskipping
This implements a frameskipping method more similar to the throttle,
which skips as many frames as necessary to catch up to real-time
deadlines, instead of just picking a constant number of frames to skip.
This reduces the reaction time in starting to skip frames, and provides
the ability to skip <1 frames per rendered frame, which eliminates the
problem of having the frameskip hover between 0 and 1 every 10 frames.

Unfortunately, whenever the sound code waits for the callback to consume
the buffer, this still results in an incorrect penalty that results in a
round of maxFrameSkip skipped frames.

The best fix for this is to reimplement the sound callback
synchronization to avoid waiting on the emulation thread. However, in
the short term, we can record the amount of time spent waiting on the
sdlBufferEmpty semaphor, and slip the deadline by that amount.